Bug Fix for Numpy 2 & Tensor Dtype &&& Add Full Negative

+++ SeqRec/trainers/sequence_trainer.py
@@ -45,7 +45,7 @@ class SeqTrainer(Trainer):
 
         for batch in prog_iter:
 
-            batch = tuple(t.to(self.device) for t in batch)
+            batch = tuple(t.long().to(self.device) for t in batch)
 
             train_start = time.time()
             inputs = self._prepare_train_inputs(batch)
@@ -94,16 +94,14 @@ class SeqTrainer(Trainer):
 
         for batch in tqdm(test_loader, desc=desc):
 
-            batch = tuple(t.to(self.device) for t in batch)
+            batch = tuple(t.long().to(self.device) for t in batch)
             inputs = self._prepare_eval_inputs(batch)
             seq_len = torch.cat([seq_len, torch.sum(inputs["seq"]>0, dim=1)])
             target_items = torch.cat([target_items, inputs["pos"]])
-            
             with torch.no_grad():
 
                 inputs["item_indices"] = torch.cat([inputs["pos"].unsqueeze(1), inputs["neg"]], dim=1)
                 pred_logits = -self.model.predict(**inputs)
-
                 per_pred_rank = torch.argsort(torch.argsort(pred_logits))[:, 0]
                 pred_rank = torch.cat([pred_rank, per_pred_rank])
